StreamnProcess/src/stream_pole_emploi/cassandra_driver/requirements.txt
api-offres-emploi==0.0.2
certifi==2023.7.22
charset-normalizer==3.3.0
cassandra-driver==3.25.0
idna==3.4
python-dateutil==2.8.2
python-decouple==3.8
pytz==2023.3.post1
six==1.16.0
tzdata==2023.3
kafka-python==2.0.2
requests==2.31.0
pandas>=1.5.3